pyproject.toml
# pyproject.toml
[build-system]
requires = ["hatchling>=1.26", "editables>=0.3"]
build-backend = "hatchling.build"

[project]
name = "vcfcache"
version = "0.2.0"
description = "Cache-based VCF annotation accelerator"
requires-python = ">=3.11"
readme = "README.md"
authors = [
    {name = "Julius Müller", email = "julius.mueller@dkfz-heidelberg.de"}
]
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "pysam>=0.21.0",
]

[project.optional-dependencies]
parquet = [
    "pandas>=2.2.0",
    "pyarrow>=18.0.0"
]
dev = [
    "pandas>=2.2.0",
    "pyarrow>=18.0.0",
    "pytest>=8.3.0", # python -m pytest -xvs
    "ruff>=0.11.9", # ruff check --fix vcfcache
    "black>=25.1.0", # black vcfcache
    "mypy>=1.15.0" # mypy vcfcache
]

[project.scripts]
vcfcache = "vcfcache.cli:main"

[tool.pytest.ini_options]
pythonpath = [
  "src"
]
testpaths = [
  "tests"
]
markers = [
  "vanilla: marks tests for vanilla package (no cache, no external tools)",
  "requires_cache: marks tests that need blueprint cache but no annotation",
  "requires_annotation: marks tests that need full annotation pipeline with VEP",
  "integration: marks tests as integration tests (legacy marker)",
  "unit: marks tests as unit tests (legacy marker)",
  "blueprint: marks tests for blueprint image (legacy marker)"
]

[tool.hatch.build]
include = [
  "vcfcache/**"
]

[tool.hatch.build.targets.wheel]
packages = ["vcfcache"]

[tool.ruff]
# Same as Black.
line-length = 88
# Exclude a variety of commonly ignored directories.
exclude = [
    ".git",
    ".mypy_cache",
    ".ruff_cache",
    ".venv",
    "__pypackages__",
    "dist",
    "build",
]

[tool.ruff.lint]
# Enable flake8-bugbear (`B`) rules.
select = ["E", "F", "B", "I", "D"]
# Exclude E501 (line too long)
ignore = ["E501", "D205", "D100", "D200", "D102", "D107", "D415", "D417"]
# Allow autofix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D"]  # Ignore docstring rules in tests

[tool.ruff.lint.mccabe]
# Unlike Flake8, default to a complexity level of 10.
max-complexity = 10

[tool.ruff.lint.pydocstyle]
convention = "google"
